MA-12504 [Android_Pi9.0]No Ethernet icon when plugin network cable.
Modify ethernet's network_score to 110 so that will always
be larger than 100 even it can not be unvalidated which will be
minused by 40(which makes it 70), and it's still larger than other
APN type's network_score.
In this way, Etherent connection will always be able to take over
wifi connection for connectivity.

Change-Id: Id73f5106d89aa9a11aeea84f2fee9eadc7f7953c
Signed-off-by: yang.tian <yang.tian@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/6011
Reviewed-by: Wang Haoran <elven.wang@nxp.com>